backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Alexey Duryagin <aiduryagih@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-binance = "^1.0.15"
strawberry-graphql = {extras = ["debug-server"], version = "^0.93.0"}
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
